shiny/requirements.txt
brotli
pandas~=2.2.2
pip-chill
plotly~=5.22.0
pysocks
rsconnect-python
ruamel.yaml
shinylive
shinywidgets~=0.3.2

shiny~=0.10.2
ultralytics~=8.2.32
pillow~=10.3.0